+++ tests/test_hosts.py
@@ -14,42 +14,42 @@ def test_index(client: FlaskClient) -> None:
     """Testing hosts.routes.index."""
     response: TestResponse = client.get("/hosts/")
     assert response.status_code == 200
-    assert b"Hosts" in response.data
-    assert b"Appearance Summary" in response.data
+    assert "Hosts" in response.text
+    assert "Appearance Summary" in response.text
 
 
 def test_appearance_summary(client: FlaskClient) -> None:
     """Testing hosts.routes.appearance_summary."""
     response: TestResponse = client.get("/hosts/appearance-summary")
     assert response.status_code == 200
-    assert b"Appearance Summary" in response.data
-    assert b"Regular Shows" in response.data
-    assert b"All Shows" in response.data
+    assert "Appearance Summary" in response.text
+    assert "Regular Shows" in response.text
+    assert "All Shows" in response.text
 
 
 def test_appearance_counts_by_year(client: FlaskClient) -> None:
     """Testing hosts.routes.appearance_counts_by_year."""
     response: TestResponse = client.get("/hosts/appearance-counts-by-year")
     assert response.status_code == 200
-    assert b"Appearance Counts by Year" in response.data
-    assert b"Regular Shows" in response.data
-    assert b"All Shows" in response.data
+    assert "Appearance Counts by Year" in response.text
+    assert "Regular Shows" in response.text
+    assert "All Shows" in response.text
 
 
 def test_appearance_counts_by_year_grid(client: FlaskClient) -> None:
     """Testing hosts.routes.appearance_counts_by_year_grid."""
     response: TestResponse = client.get("/hosts/appearance-counts-by-year/grid")
     assert response.status_code == 200
-    assert b"Appearance Counts by Year: Grid" in response.data
-    assert b"Total" in response.data
+    assert "Appearance Counts by Year: Grid" in response.text
+    assert "Total" in response.text
 
 
 def test_appearances_by_year(client: FlaskClient) -> None:
     """Testing hosts.routes.appearances_by_year."""
     response: TestResponse = client.get("/hosts/appearances-by-year")
     assert response.status_code == 200
-    assert b"Appearances by Year" in response.data
-    assert b"Select a Host" in response.data
+    assert "Appearances by Year" in response.text
+    assert "Select a Host" in response.text
 
 
 @pytest.mark.parametrize("host_slug", ["peter-sagal", "luke-burbank"])
@@ -59,22 +59,22 @@ def test_appearances_by_year_post(client: FlaskClient, host_slug: str) -> None:
         "/hosts/appearances-by-year", data={"host": host_slug}
     )
     assert response.status_code == 200
-    assert b"Appearances by Year" in response.data
+    assert "Appearances by Year" in response.text
     assert host_slug.encode("utf-8") in response.data
-    assert b"Repeat Of" in response.data
+    assert "Repeat Of" in response.text
 
 
 def test_debuts_by_year(client: FlaskClient) -> None:
     """Testing hosts.routes.appearances_by_year_grid."""
     response: TestResponse = client.get("/hosts/debuts-by-year")
     assert response.status_code == 200
-    assert b"Debuts by Year" in response.data
-    assert b"# of Regular Appearances" in response.data
+    assert "Debuts by Year" in response.text
+    assert "# of Regular Appearances" in response.text
 
 
 def test_guest_host_counts_by_year(client: FlaskClient) -> None:
     """Testing hosts.routes.guest_host_counts_by_year."""
     response: TestResponse = client.get("/hosts/guest-host-counts-by-year")
     assert response.status_code == 200
-    assert b"Guest Host Appearance Counts by Year" in response.data
-    assert b'<th scope="col">Year</th>' in response.data
+    assert "Guest Host Appearance Counts by Year" in response.text
+    assert '<th scope="col">Year</th>' in response.text