+++ tests/test_on_this_day.py
@@ -5,17 +5,52 @@
 # vim: set noai syntax=python ts=4 sw=4:
 """Testing On This Day Routes Module and Blueprint Views."""
 
+import pytest
 from flask import Flask
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
 
 def test_index(app: Flask, client: FlaskClient):
+    """Testing on_this_day.routes.index."""
     _report_enabled: bool = app.config["app_settings"].get(
         "enable_on_this_day_report", False
     )
     response: TestResponse = client.get("/on-this-day")
     if _report_enabled:
         assert response.status_code == 200
+        assert "omnibus report" in response.text
+        assert "Host Debuts" in response.text
+        assert "Panelist Debuts" in response.text
+    else:
+        assert response.status_code == 404
+
+
+@pytest.mark.parametrize("month, day", [(1, 3), (9, 21)])
+def test_month_day(app: Flask, client: FlaskClient, month: int, day: int):
+    """Testing on_this_day.routes.month_day."""
+    _report_enabled: bool = app.config["app_settings"].get(
+        "enable_on_this_day_report", False
+    )
+    response: TestResponse = client.get(f"/on-this-day/{month}/{day}")
+    if _report_enabled:
+        assert response.status_code == 200
+        assert "omnibus report" in response.text
+        assert "Host Debuts" in response.text
+        assert "Panelist Debuts" in response.text
+    else:
+        assert response.status_code == 404
+
+
+@pytest.mark.parametrize("month, day", [(13, 3), (9, 32)])
+def test_month_day_invalid(app: Flask, client: FlaskClient, month: int, day: int):
+    """Testing on_this_day.routes.month_day."""
+    _report_enabled: bool = app.config["app_settings"].get(
+        "enable_on_this_day_report", False
+    )
+    response: TestResponse = client.get(f"/on-this-day/{month}/{day}")
+    if _report_enabled:
+        assert response.status_code == 302
+        assert "on-this-day" in response.location
     else:
         assert response.status_code == 404